requirements.txt
Pillow
numpy
gradio